third_party/blink/web_tests/fast/dom/shadow/selection-shouldnt-expose-shadow-dom-expected.txt

CONSOLE WARNING: line 27: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS internals.treeScopeRootNode(selection.anchorNode) is document
PASS internals.treeScopeRootNode(selection.focusNode) is document
PASS successfullyParsed is true

TEST COMPLETE
